Find and Download unsaved Images and Videos
Query firebase to find all events that have not been backed up to S3, then download each file. We want to keep disk and memory usage low, so each file is streamed and writen to disk without storing the entire file in memory. The next step is to save each file to S3.

- created an iter to get events that need to be backed_up from firebase
- created file_utils to manage downloading files